Remove mainhelper methods and add _label
