+++ com_migratetojoomla/admin/language/en-GB/com_migratetojoomla.ini
@@ -7,8 +7,8 @@ COM_MIGRATETOJOOMLA="Migrate To Joomla"
 COM_MIGRATETOJOOMLA_CHECK_MEDIA_CONNECTION="Check media connection"
 COM_MIGRATETOJOOMLA_DATABASE_CONNECTION_SUCCESSFULLY="Database connection successfully"
 COM_MIGRATETOJOOMLA_DATABASE_CONNECTION_UNSUCCESSFULLY="Database connection unsuccessfully"
-COM_MIGRATETOJOOMLA_DATABASE_DRIVER="Database Driver "
-COM_MIGRATETOJOOMLA_DATABASE_NAME="Database Name"
+COM_MIGRATETOJOOMLA_DATABASE_DRIVER_LABEL="Database Driver "
+COM_MIGRATETOJOOMLA_DATABASE_NAME_LABEL="Database Name"
 COM_MIGRATETOJOOMLA_DOWNLOAD_MEDIA="Download Media"
 COM_MIGRATETOJOOMLA_DOWNLOAD_MEDIA_SUCCESSFULLY="Download media successfully"
 COM_MIGRATETOJOOMLA_DOWNLOAD_MEDIA_UNSUCCESSFULLY="Download media unsuccessfully"
@@ -16,28 +16,28 @@ COM_MIGRATETOJOOMLA_FS="File System"
 COM_MIGRATETOJOOMLA_FS_CONNECTION_SUCCESSFULLY="File system Connection Successfully"
 COM_MIGRATETOJOOMLA_FS_CONNECTION_UNSUCCESSFULLY="File system Connection Unuccessfully"
 COM_MIGRATETOJOOMLA_FTP="FTP"
-COM_MIGRATETOJOOMLA_FTP_BASE_DIR="FTP base directory"
+COM_MIGRATETOJOOMLA_FTP_BASE_DIR_LABEL="FTP base directory"
 COM_MIGRATETOJOOMLA_FTP_CONNECTION_SUCCESFULLY="Ftp Connection Successfully"
 COM_MIGRATETOJOOMLA_FTP_CONNECTION_UNSUCCESFULLY="Ftp Connection Unsuccessfully"
-COM_MIGRATETOJOOMLA_FTP_HOST="FTP host"
-COM_MIGRATETOJOOMLA_FTP_PASSWORD="FTP password"
-COM_MIGRATETOJOOMLA_FTP_PORT="FTP port"
-COM_MIGRATETOJOOMLA_FTP_PROTOCOL="FTP protocol"
-COM_MIGRATETOJOOMLA_FTP_USERNAME="FTP username"
+COM_MIGRATETOJOOMLA_FTP_HOST_LABEL="FTP host"
+COM_MIGRATETOJOOMLA_FTP_PASSWORD_LABEL="FTP password"
+COM_MIGRATETOJOOMLA_FTP_PORT_LABEL="FTP port"
+COM_MIGRATETOJOOMLA_FTP_PROTOCOL_LABEL="FTP protocol"
+COM_MIGRATETOJOOMLA_FTP_USERNAME_LABEL="FTP username"
 COM_MIGRATETOJOOMLA_FTPS="FTPS"
-COM_MIGRATETOJOOMLA_HOSTNAME="Hostname"
+COM_MIGRATETOJOOMLA_HOSTNAME_LABEL="Hostname"
 COM_MIGRATETOJOOMLA_HTTP="Http"
 COM_MIGRATETOJOOMLA_HTTP_CONNECTION_SUCCESSFULLY="Http Connection Successfully"
 COM_MIGRATETOJOOMLA_HTTP_CONNECTION_UNSUCCESSFULLY="Http Connection Unsuccessfuly"
-COM_MIGRATETOJOOMLA_HTTP_LINK="Wordpress live website URL"
-COM_MIGRATETOJOOMLA_PASSWORD="Password"
-COM_MIGRATETOJOOMLA_PORT="Port"
-COM_MIGRATETOJOOMLA_SELECT_OPTION_FOR_MEDIA="Select option for media"
+COM_MIGRATETOJOOMLA_HTTP_LINK_LABEL="Wordpress live website URL"
+COM_MIGRATETOJOOMLA_PASSWORD_LABEL="Password"
+COM_MIGRATETOJOOMLA_PORT_LABEL="Port"
+COM_MIGRATETOJOOMLA_SELECT_OPTION_FOR_MEDIA_LABEL="Select option for media"
 COM_MIGRATETOJOOMLA_SFTP="SFTP"
 COM_MIGRATETOJOOMLA_TEST_DATABASE_CONNECTION="Test Database Connection"
-COM_MIGRATETOJOOMLA_TABLE_PREFIX="Table prefix"
-COM_MIGRATETOJOOMLA_USERNAME="Username"
-COM_MIGRATETOJOOMLA_WEBSITE_BASE_DIRECTORY="Website base directory"
+COM_MIGRATETOJOOMLA_TABLE_PREFIX_LABEL="Table prefix"
+COM_MIGRATETOJOOMLA_USERNAME_LABEL="Username"
+COM_MIGRATETOJOOMLA_WEBSITE_BASE_DIRECTORY_LABEL="Website base directory"
 COM_MIGRATETOJOOMLA_WORDPRESS_FTP_PARAMETERS="Wordpress ftp parameters"
 COM_MIGRATETOJOOMLA_WORDPRESS_DATABASE_PARAMETERS="Wordpress Database Parameters"
 COM_MIGRATETOJOOMLA_WORDPRESS_WEBSITE_PARAMETERS="Wordpress Website Parameters"
