+++ com_migratetojoomla/admin/src/controller/DisplayController.php
@@ -1,11 +1,5 @@
 <?php
 
-namespace Joomla\Component\MigrateToJoomla\Administrator\Controller;
-
-defined('_JEXEC') or die;
-
-use Joomla\CMS\MVC\Controller\BaseController;
-
 /**
  * @package     Joomla.Administrator
  * @subpackage  com_migratetojoomla
@@ -13,6 +7,14 @@ use Joomla\CMS\MVC\Controller\BaseController;
  * @copyright     (C) 2023 Open Source Matters, Inc. <https://www.joomla.org>
  * @license       GNU General Public License version 2 or later; see LICENSE.txt
  */
+
+namespace Joomla\Component\MigrateToJoomla\Administrator\Controller;
+
+// phpcs:disable PSR1.Files.SideEffects
+\defined('_JEXEC') or die;
+// phpcs:enable PSR1.Files.SideEffects
+
+use Joomla\CMS\MVC\Controller\BaseController;
 
 /**
  * Default Controller of Migrate To Joomla component
